KS-10-4 remove v1. review fixes
